Register user ready

+++ app/core/config.py
@@ -47,8 +47,7 @@ class AuthJWT(BaseModel):
     private_key_path: Path = BASE_DIR / "jwt-private.pem"
     public_key_path: Path = BASE_DIR / "jwt-public.pem"
     algorithm: str = "RS256"
-    lifetime_seconds: int = 3600
-
+    lifetime_seconds: float = 3600.0
 
 class Settings(BaseSettings):
     model_config = SettingsConfigDict(
